cleanup the AST API
