+++ apollo-compiler/src/test/kotlin/com/apollographql/apollo3/compiler/ValidationTest.kt
@@ -3,7 +3,7 @@ package com.apollographql.apollo3.compiler
 import com.apollographql.apollo3.ast.Issue
 import com.apollographql.apollo3.ast.ParseResult
 import com.apollographql.apollo3.ast.parseAsGQLDocument
-import com.apollographql.apollo3.ast.validateAsOperations
+import com.apollographql.apollo3.ast.validateAsExecutable
 import com.apollographql.apollo3.ast.validateAsSchema
 import com.apollographql.apollo3.compiler.TestUtils.checkExpected
 import com.apollographql.apollo3.compiler.TestUtils.testParametersForGraphQLFilesIn
@@ -29,7 +29,7 @@ class ValidationTest(name: String, private val graphQLFile: File) {
 
       when (parseResult) {
         is ParseResult.Error -> parseResult.issues
-        is ParseResult.Success -> parseResult.value.validateAsOperations(schema!!)
+        is ParseResult.Success -> parseResult.value.validateAsExecutable(schema!!)
       }
     } else {
       val parseResult = graphQLFile.parseAsGQLDocument()
