--- apollo-graphql-ast/src/main/kotlin/com/apollographql/apollo3/ast/gqlfielddefinition.kt
@@ -1,5 +1,0 @@
-package com.apollographql.apollo3.ast
-
-internal fun GQLFieldDefinition.isDeprecated(): Boolean {
-  return directives.firstOrNull { it.name == "deprecated" } != null
-}
