+++ apollo-graphql-ast/src/main/kotlin/com/apollographql/apollo3/ast/gqlfragmentdefinition.kt
@@ -1,14 +1,6 @@
 package com.apollographql.apollo3.ast
 
+import com.apollographql.apollo3.ast.internal.ExecutableValidationScope
 
 
-fun GQLFragmentDefinition.validate(
-    schema: Schema,
-    fragments: Map<String, GQLFragmentDefinition>,
-) = ExecutableValidationScope(schema, fragments).validateFragment(this)
-
-fun GQLFragmentDefinition.inferVariables(
-    schema: Schema,
-    fragments: Map<String, GQLFragmentDefinition>,
-) = ExecutableValidationScope(schema, fragments).inferFragmentVariables(this)
