+++ apollo-graphql-ast/src/main/kotlin/com/apollographql/apollo3/ast/gqloperationdefinition.kt
@@ -1,5 +1,7 @@
 package com.apollographql.apollo3.ast
 
+import com.apollographql.apollo3.ast.internal.ExecutableValidationScope
+
 fun GQLOperationDefinition.rootTypeDefinition(schema: Schema) = when (operationType) {
   "query" -> schema.queryTypeDefinition
   "mutation" -> schema.mutationTypeDefinition
