+++ apollo-graphql-ast/src/main/kotlin/com/apollographql/apollo3/ast/gqltypedefinition.kt
@@ -21,7 +21,7 @@ fun GQLTypeDefinition.isFieldNonNull(fieldName: String): Boolean {
 
   val stringValue = (directive.arguments!!.arguments.first().value as GQLStringValue).value
 
-  val selections = stringValue.parseAsSelections().getOrNull() ?: error("'$stringValue' is not a valid selectionSet")
+  val selections = stringValue.parseAsGQLSelections().getOrNull() ?: error("'$stringValue' is not a valid selectionSet")
 
   return selections.filterIsInstance<GQLField>()
       .map { it.name }
