+++ apollo-graphql-ast/src/main/kotlin/com/apollographql/apollo3/ast/internal/SchemaValidationScope.kt
@@ -1,4 +1,6 @@
-package com.apollographql.apollo3.ast
+package com.apollographql.apollo3.ast.internal
+
+import com.apollographql.apollo3.ast.*
 
 /**
  */
